Fold my layout experiments back into this repo.
(The other repo will now be deleted). In the end, this involved moving the LibraryResources directory into gitLInk, so that we can use that gitLink directory in a pinch for ad hoc distribution.

Also change the name of the context to gitLink`, to match common usage. (Now we'll need to decide whether all the exported symbols should be lowercased too.)