More pacletization work.
Add code to assemble and build new .paclet files for deployment. I have switched to a FileTemplate-based PacletInfo creation system, which lets us synthesize version numbers based on the current date and time.

This assembly code will probably be superceded by RE build procedures, when the time comes. But for now, it's a big time saver.