C++ implementation of GitCommit.
Untested, and no WL implementation yet.  But it's now documented and
mostly implemented, and it compiles.  Main thing not done on the C++ side
is figuring out how to deal with the signature metadata.

Also, lots of refactoring, and adding some new functionality which will
be required in the not-so-distant future.