Implement GitCheckout and GitType.
This was more painful than anticipated...

* Begin the transition to using GitObject.  More work to
do there.
* Implement GitCheckout.  Which, internally, breaks
down into GitSetHead and GitCheckoutHead.  But
also requires the ability to create brand new branches
and set their upstream branches apprropriately.
So more than the usual WL implementation here.
* Implemented "CheckoutStrategy" option for
GitCheckout.  Which should basically do as libgit2
documents, but either some of the documentation
is off or something is buggy because it's not quite
behaving as I expect.  "CheckoutStrategy"->{"Force"}
seem to be the way to go generally if you don't mind
throwing away things.  The "Safe" and "SafeCreate"
leave a lot more litter behind than expected.
* Begin, but not finish, implementing "Notifications"
option.
* Add MLExpr infrastructure for better handling the
"CheckoutStrategy" and "Notifications" options.