Add GitMoveBranch, and make GitMerge use it.
Now GitMerge will move the branch you're committing to.  If
the branch you're committing to is the HEAD branch (or if you
literally used HEAD), then it will do a GitCheckout as well to
update things.  Note that this GitCheckout uses
"CheckoutStrategy"->{"Force"} right now, which is too
destructive, but anything more nuanced will require a lot of work.