Implement GitInit
This also fixes a problem Itai reported with GitClone
where it was, by default, creating world-writable
directories.  Turns out this is not a lever you can
control in libgit2's behavior for cloning, but it
*is* one you can control for git-init, and git-clone
is quite happy not having to create the directory.

So I have a bit of a hack whereby GitClone calls
GitInit, then deletes the .git directory.