Infrastructure toward git tree diffing.
The idea is to get a list of files which are different between any two given
trees.  In preparation for GitCheckoutReference, which needs to know
this information to check for conflicts with the index and working tree.

Everything compiles, and lots of infrastructure in place, but the actual
diffs aren't yet being returned.