Fix GitCheckoutReference[] for filemode issues.
GitCheckoutReference would sometimes fail when a file had a filemode
change.  This happened on all platforms.  Because I was setting the
ref head first, libgit2 didn't have two stable git trees to compare, so
it had to do all of its comparisons with the working tree.  Turns out that
working tree comparisons are not as robust as one might hope.

Compound that difficulty with the fact that Windows can't do Unix filemodes,
and now you have real problems.  But libgit2 doesn't even do filemodes from
the working tree correctly on other platforms, so it's a mixed bag.

So, now I check out the target tree first, then set HEAD if that succeeds.

Also, add a few extra error-checks, fix some consts, etc.