Make GitCommit support creating explicitly orphaned commits.
I.e., GitCommit[repo, log, tree, None]

I'll probably have to fix this a bit more when I support GitAdd
to deal with the very first commit in a repo, which should work
without having to specify None explicitly.

Also, catch up on documentation updates with the state of the code.