Implement GitAdd, GitReset.
What a major pain in the butt.  libgit2 won't even lift the
tiniest finger to help me here.  If I want to add an explicit
path which doesn't exist on disk (thereby adding the deletion),
I have to do it myself.  If I want to implement pretty much anything
that "git reset <filename>" does, I have to do it myself.  Jeez.

So, this is implemented in terms of my previous work.  Use
the RepoStatus class to figure out which files are in the working
path and which ones are in the index, and deal with them appropriately.
Use the GitTree class to figure out how to recreate index entries
equivalent to the tree entries in HEAD.

Make the tests do a better job of cleaning up after themselves, so I
don't end up with cross-contamination between tests.