Make GitRepo, GitObject durable.
* Remove all of the managed instance stuff.
* Replace GitRepo contents with an association which
contains the elements of GitProperties[_GitRepo] which
cannot change through normal git operations.
* Make GitRepo take DownValues which do GitProperties
lookups.  For the fields in the association, get the properties
straight from there...otherwise bail to GitProperties.
* Everything internally that used the pattern
   f[GitRepo[id_]] := GL`foo[id]
now does instead...
   f[repo_GitRepo] := GL`foo[repo["GitDirectory"]]
* Remove the remaining outlier GitLink functions which
weren't using MLINKs to communicate with the kernel.
* GitClose is now a real thing.  It may be a bit hard to
tell that it works because of the durability of GitRepo
objects, but this should allow Windows to release its
file locks on the git repo.

And unrelated, but while I was rooting around in the
code, I realized that the canonizePath stuff I added
for GitAdd/Reset to fix up pathname separators was
entirely unnecessary because I already had a C++
solution for this (clever me).  So, implement that C++
solution for GitStatus, GitAdd, and GitReset.