*Now*, GitClose is working correctly.
Tested under Windows and in a debugger.  Fixed some
cases of duplicate repos being opened (and leaked) behind
the scenes.  Now, if you close a repo, you really can delete it.
Well, not from within Mathematica because of limitations in
DeleteDirectory, but at least from within a shell.

Also, discovered that GitAdd/GitReset were behaving funny
on Windows because of inconsistencies in pathname strings.
Added a new type for pathnames and imposed it everywhere
to try to make sense out of this (i.e. because libgit insists on
using / while we want to use \ on Windows).

And also GitAdd/Reset could mess up the encodings of
returned pathnames.  Now fixed.