Fix up GitRepo object.
There were a couple of things in the association
that shouldn't have been.  One of them seems to have
exposed a libgit2 bug which I spent a while tracing,
but I'm giving up for the time being.

I.e., when I made hundreds of thousands of calls to
git_repository_is_empty(), it either hands or takes
such an absurdly long time as to appear to hang.
The hang is due to the internal use of
git_reference_iterator_new(), which looks like it
ought to be cached, pooled, etc.  But perhaps
not.

But it shouldn't have been in the association anyway.
The only things that should be in that association
are things that would require fundamental surgery
to the git repo to change (like whether or not it's
bare, its working directory, etc.).