Make it compile against libgit v0.23.4
Turn on some new-to-v0.23 features, mostly choices having
to do with fetch and merge.  But merge seems busted for the
moment.

Fixed Options[GitPull], which now merges Options[GitMerge] and
Options[GitFetch].

Had to rewrite GitSignature[] a bit...in the new libgit2,
git_config_get_string() only works on a git config snapshot now.
That sounds expensive to make, especially for an operation which
is done *all* the time, so I switched to using git_config_get_string_buf().