Fix authentication.  Again.
Seems that, internally, the ssh authentication can loop.
Quite a lot.  This seems like a bug in libgit2's use of
ssh authentication.  Or an utter lack of documentation.
As they just keep calling our stuff and we keep providing
the same credentials (not sure if any non-ssh authenticators
do this or not).

In any case, rewrote the RemoteConnector so that
it's generally going to be a stack-allocated object,
and is robust to having its credentials callbacks called
multiple times by libgit2, and this is how it downgrades
from ssh-agent authentication to file-level authentication.
And if it gets called again, it downgrades to *failing* the
credentials callback.  Which prevents the callbacks from
being called what might be zillions of times.

It's still trying ssh-agent authentication first, which seems
wrong, as doing so actually *launches* ssh-agent (at least
on Mac) even if it's not running.  But libssh2 doesn't
appear to have a way to detect whether ssh-agent is
running or not.

The above was on Mac (and probalby Linux)...on Windows it fails
to launch ssh-agent and bails out normally.  So, two code paths
here for the different platforms.