Dramatically speed up GitStatus[].
Turns out that GitStatus[] supports a stat cache.  But, for reasons
unknown, it is not turned on by default.  Turning this on speeds
up the population of the up/down/status bubbles in Stashboard
dramatically (in my case, by a factor of three)...on the second call.
The first call is still slow as it has to build the cache.