Add SystemID and Qualifier to PacletInfoTemplate.m.
...since we are building platform-specific paclets & need to tell the paclet server about it.

The Qualifier & SystemID options are undocumented, but:

Qualifier can be any arbitrary string. Here, we use the platform code.

SystemID is a list of $SystemID values. For now, the list is hard-coded to contain a single value. If/when somebody wants a Win32+Win64 paclet, that's going to cause problems.