Build both SSL 1.1 and SSL 1.0 versions of GitLink + cleanup
Because SSL 1.1 is incompatible with versions 12.0 or lower of the kernel,
which uses SSL 1.0, we need to build two versions of the GitLink in order
to have compatibility with all version from 11.1+.  At runtime, the
correct library is loaded based on the kernel version.  At this point,
only the developer build-script is updated.

Note: when building libgit2 on Windows, it is important to build with its
internal copy of zlib.  Otherwise it is necessary to add an explicit link
dependency on zlib for GitLink.

While we're at it, rip out the code to load gitLink_10_3 on OS X, which
we haven't built in a while and is irrelevant now in a world where the
paclet is 11.1+.  Note that the OS X deployment target is set differently
on the two GitLink .dylib's to reflect how support libraries were built
(in order to match the corresponding kernel versions).