Add explicit library directories for OpenSSL on OS X.
Libgit2 on OSX has a dependency on OpenSSL, same as Linux, as evidenced
by the "crypto" link library it needs.  Back in the way old days, we got
it from the OSX SDK, but it was removed from Sierra's SDK.  Since then,
we've been getting it from SystemFiles/Libraries/MacOSX-x86-64, whether
consciously or not, because CCompilerDriver automatically adds it as a
link directory.  Since we're now trying to link against two different
versions of OpenSSL (to support multiple M versions), it seems prudent
to have explicit link directories to make sure each version links
against a matching library.

While I was at it, updated the Linux SSL 1.0.x version to 1.0.2s,
matching the version on OS X.