Future proof Needs statement for planned changes to the DocumentationBuild "prelude"
Ideally Needs["DocumentationBuild`"] wouldn't result in a dozen internal/private
contexts ending up on $ContextPath, but at the moment that's the case. We'd like
to change that in the future, so I'm future-proofing this particular Needs
statement against that, by loading the precise context which is used in this code.